倍率设置
全部完成，编译通过 `BUILD SUCCESS`。

## 改动总结

### 配置入口
**`resources/server.properties`** — 所有倍率/几率在此定义，修改后重启生效，**无需重新编译**。

### 代码变化
| 文件 | 迁移到 ServerConfig 的常数量 |
|---|---|
| `GameConstants.java` | 42 个（经验/掉率/枫币/NX/刷新/符文等） |
| `CustomConstants.java` | 5 个（强化怪HP/伤害/水晶价格/税率） |
| `ItemConstants.java` | 8 个（潜能/星岩/火焰系数跳过数组） |
| `BossConstants.java` | 3 个（Lotus落脚/HT混沌/巨花掉落） |
| `OzConstants.java` | 6 个（勋章掉率/史莱姆倍率） |
| `SkillChangeConstants.java` | 11 个（技能触发率/持续时间等） |
| `ServerConstants.java` | 1 个（缓存清理周期） |
| `EliteChampionFieldEvent.java` | 2 个 |
| `MoonBunnyInstanceHelper.java` | 1 个 |

### 使用方式

```properties
# resources/server.properties
server.expMultiplier = 500       # 经验倍率
server.mobDropRate = 1           # 掉率
server.mobMesoRate = 2           # 枫币倍率
server.mobNxRate = 1             # NX 倍率
# ... 总共 ~79 个配置项
```

**未迁移的**（数组类型，properties 不便表达）：
- `ItemConstants.WEAPON_FLAME_MULTIPLIER[]`
- `ItemConstants.WEAPON_FLAME_MULTIPLIER_BOSS_WEAPON[]`
- `BossConstants.GOLLUX_HP_MULTIPLIERS[][]`

+++ src/main/java/net/swordie/ms/client/character/Char.java
@@ -382,7 +382,7 @@ public class Char {
     private Set<FirstEnterReward> firstEnterRewards;
     private MiniRoom miniRoom;
     private List<MiniGameRecord> miniGameRecords;
-    private int currentExpRate;
+    private double currentExpRate;
 
     public Char() {
         itemPots = new ArrayList<>();
@@ -2668,7 +2668,7 @@ public class Char {
         }
     }
 
-    public int getCurrentLevelExpRate() {
+    public double getCurrentLevelExpRate() {
         if (currentExpRate == 0) {
             updateExpRate();
         }
